+++ src/routes/router.tsx
@@ -2,12 +2,15 @@ import React from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 
 import { ChartPage as HomePage } from '../pages/ChartPage';
-import { HomePageRouteErrorBoundary } from '../components/HomePageRouteErrorBoundary';
+import { ErrorPage } from '../pages/ErrorPage';
 
 export const router = createBrowserRouter([
   {
     path: '/',
     element: <HomePage />,
-    errorElement: <HomePageRouteErrorBoundary />,
+  },
+  {
+    path: '*',
+    element: <ErrorPage />,
   },
 ]);
